temp 4: create keyring module
